Factor BufferLayerManager out to separate file
