+++ browser/src/Editor/NeovimEditor/NeovimEditor.tsx
@@ -68,6 +68,7 @@ import { normalizePath, sleep } from "./../../Utility"
 
 import * as VimConfigurationSynchronizer from "./../../Services/VimConfigurationSynchronizer"
 
+import { BufferLayerManager } from "./BufferLayerManager"
 import { Definition } from "./Definition"
 import * as ActionCreators from "./NeovimEditorActions"
 import { NeovimEditorCommands } from "./NeovimEditorCommands"
@@ -75,63 +76,6 @@ import { createStore, IState } from "./NeovimEditorStore"
 import { Rename } from "./Rename"
 import { Symbols } from "./Symbols"
 import { IToolTipsProvider, NeovimEditorToolTipsProvider } from "./ToolTipsProvider"
-
-export type BufferLayerFactory = (buf: Oni.Buffer) => Oni.EditorLayer
-
-export type BufferFilter = (buf: Oni.Buffer) => boolean
-
-export const createBufferFilterFromLanguage = (language: string) => (buf: Oni.Buffer): boolean => {
-    if (!language || language === "*") {
-        return true
-    } else {
-        return buf.language === language
-    }
-
-}
-
-export interface BufferLayerInfo {
-    filter: BufferFilter
-    layerFactory: BufferLayerFactory
-}
-
-export class BufferLayerManager {
-    private _layers: BufferLayerInfo[] = []
-
-    private _buffers: Oni.Buffer[] = []
-    
-    public addBufferLayer(filterOrLanguage: BufferFilter | string, layerFactory: BufferLayerFactory) {
-
-        let filter: BufferFilter
-        if (typeof filterOrLanguage === "string") {
-            filter = createBufferFilterFromLanguage(filterOrLanguage)
-        } else {
-            filter = filterOrLanguage
-        }
-
-        this._layers.push({
-            filter,
-            layerFactory,
-        })
-
-        this._buffers.forEach((buf) => {
-            if (filter(buf)) {
-                buf.addLayer(layerFactory(buf))
-            }
-        })
-    }
-
-    public notifyBufferEnter(buf: Oni.Buffer): void {
-        if (this._buffers.indexOf(buf) === -1) {
-            this._buffers.push(buf)
-
-            this._layers.forEach((layerInfo) => {
-                if (layerInfo.filter(buf)) {
-                    buf.addLayer(layerInfo.layerFactory(buf))
-                }
-            })
-        }
-    }
-}
 
 import { BufferScrollBarContainer } from "./containers/BufferScrollBarContainer"
 import { DefinitionContainer } from "./containers/DefinitionContainer"
